chatbot/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.5
django-extensions==3.2.3
django-filter==24.2
django-unfold==0.32.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
firebase-admin==6.5.0
frozenlist==1.4.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
msgpack==1.1.0
multidict==6.1.0
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.9.0
pyparsing==3.1.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.13.1
